requirements.txt
# ── PDF & OCR ──────────────────────────────────────────────
pypdf>=4.1.0
pdfplumber>=0.10.3
pytesseract>=0.3.10
pdf2image>=1.17.0
Pillow>=10.2.0

# ── Banco de Dados ─────────────────────────────────────────
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.28

# ── Notificações ───────────────────────────────────────────
python-telegram-bot>=21.0.1
slack-sdk>=3.27.0

# ── LLM Local (Ollama) ─────────────────────────────────────
ollama>=0.1.8
httpx>=0.27.0

# ── Utilitários ────────────────────────────────────────────
watchdog>=4.0.0
python-dotenv>=1.0.1
pydantic>=2.6.4
pydantic-settings>=2.2.1
structlog>=24.1.0
rich>=13.7.1
jinja2>=3.1.3
